fix: close R2 verification regressions
