+++ packages/apps/src/ditto_apps/scripts/__init__.py
@@ -1,3 +1,5 @@
 """Executable operational acceptance helpers."""
 
 from __future__ import annotations
+
+__all__: list[str] = []
