+++ packages/apps/tests/integration/test_main_routes_integration.py
@@ -66,6 +66,7 @@ class TestMainRoutesAssembly:
             "backtest",
             "capital",
             "commodity",
+            "data_products",
             "fundamental",
             "fx",
             "ingestion",
