+++ packages/apps/tests/unit/registry/test_ingestion_context_unit.py
@@ -6,10 +6,15 @@ from contextlib import contextmanager
 from typing import cast
 from unittest.mock import MagicMock
 
+from ditto_application.commands.data_product_certification import (
+    DataProductCertificationCommands,
+)
+from ditto_application.processes.ingestion.bootstrap_planner import BootstrapPlanner
 from ditto_application.processes.ingestion.coordinator_factory import (
     CoordinatorRuntimeContext,
     CoordinatorServices,
 )
+from ditto_application.queries.data_products import DataProductsQueryFacade
 from ditto_apps.registry.contexts import ingestion as ingestion_context
 from ditto_data.catalog import (
     DataCatalogReader,
@@ -17,7 +22,6 @@ from ditto_data.catalog import (
     InMemoryDataCatalog,
 )
 from ditto_data.catalog.fallback_policy import CatalogSourceFallbackPolicyReader
-from ditto_data.ingestion.partition_state import PartitionLifecycleReader
 from ditto_data.lineage import DataLineageRecorder, InMemoryDataLineage
 from ditto_data.sources.registry import SourceRegistry
 
@@ -58,7 +62,9 @@ def test_create_ingestion_bundle_passes_lineage_recorder(mocker) -> None:
         DataCatalogReader: catalog,
         DataCatalogWriter: catalog,
         CatalogSourceFallbackPolicyReader: source_fallback_policy_reader,
-        PartitionLifecycleReader: MagicMock(),
+        BootstrapPlanner: MagicMock(),
+        DataProductsQueryFacade: MagicMock(),
+        DataProductCertificationCommands: MagicMock(),
     }
     container = _FakeContainer(services)
     coordinator = MagicMock()
