+++ packages/data/src/ditto_data/catalog/__init__.py
@@ -7,22 +7,9 @@ from ditto_data.catalog.contracts import (
     DataCatalogWriter,
     DataSchemaFingerprint,
 )
-from ditto_data.catalog.license import (
-    DatasetLicenseDraft,
-    DatasetLicenseReader,
-    DatasetLicenseRecord,
-    DatasetLicenseWriter,
-)
 from ditto_data.catalog.metadata import (
     DatasetMetadata,
-    DatasetProductContract,
     default_dataset_metadata,
-)
-from ditto_data.catalog.source_snapshot import (
-    ProviderSnapshot,
-    ProviderSnapshotDraft,
-    ProviderSnapshotReader,
-    ProviderSnapshotWriter,
 )
 from ditto_data.catalog.store import InMemoryDataCatalog
 
@@ -32,16 +19,7 @@ __all__ = [
     "DataCatalogReader",
     "DataCatalogWriter",
     "DataSchemaFingerprint",
-    "DatasetLicenseDraft",
-    "DatasetLicenseReader",
-    "DatasetLicenseRecord",
-    "DatasetLicenseWriter",
     "DatasetMetadata",
-    "DatasetProductContract",
     "InMemoryDataCatalog",
-    "ProviderSnapshot",
-    "ProviderSnapshotDraft",
-    "ProviderSnapshotReader",
-    "ProviderSnapshotWriter",
     "default_dataset_metadata",
 ]
